+++ mall-portal/src/main/java/com/macro/mall/portal/config/MallSecurityConfig.java
@@ -1,0 +1,29 @@
+package com.macro.mall.portal.config;
+
+import com.macro.mall.portal.service.UmsMemberService;
+import com.macro.mall.security.config.SecurityConfig;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+/**
+ * mall-security模块相关配置
+ * Created by macro on 2019/11/5.
+ */
+@Configuration
+@EnableWebSecurity
+@EnableGlobalMethodSecurity(prePostEnabled=true)
+public class MallSecurityConfig extends SecurityConfig {
+
+    @Autowired
+    private UmsMemberService memberService;
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        //获取登录用户信息
+        return username -> memberService.loadUserByUsername(username);
+    }
+}
